Bug fix: row_header height
When number of rows extended past the height of the frame, the row
headers would increase in height and no longer match up with the rows.
The more rows, the taller each row header. This change makes the row
headers the same height as the rows no matter how many rows there are.