Work on data updating in adapters and ListView
The adapter and ListView system has several open issues about data
updating, and ListView's reaction for scrolling. From discussion with
tito and reference to SproutCore, more information needs to be added
about the nature of data change, which will be different for
ListProperty and DictProperty.

So far, for lists, a new cls parameter is added to ListProperty, so
that instead of using ObservableList always, it will check to see if
an ObservableList variant is provided. In this case cls will be set
to a new RangeObservingList for ListAdapter. Something similar will
be needed for DictAdapter.

As this branch has been open locally for a while, several discussions
with quanon about data updating resulted in a fix for how ListView
handles its adapter binding, to assure proper bindings to ListView
updating methods are reset when adapter changes.

Also, in the same area of ListView, the bind_triggers_to_view method
has been renamed to data_changed.