Added a start-stop system; worked on list/dict ops
Added a stop/start system for the change monitor.

Broke code that responds to data changes out of ListAdapter, and into a
new ListOpHandler. Also using this for sorted_keys in DictAdapter

The data DictProperty in DictAdapter is now observed for changes in the
same way the data list (ListProperty) is observed in ListAdapter.

Improved the way dict changes are handled in DictAdapter, either
dispatching for a dict specific change, or by changing sorted_keys.

Improved scrolling reaction to item adds.

Conflicts:
	kivy/uix/listview.py