Improved ListView components and selection
Removed ListItemLabel, because Label will suffice.

SelectableView now subclasses ButtonBehavior, so gets a proper
on_release, and the trigger_action() method is also used. A
carry_selection_to_children boolean was added, so that in the select()
and deselect() methods the equivalent methods on children can be
optionally called.

CompositeListItem lost its is_representing_cls, which was mostly for
debugging. It gained a bind_selection_from_children boolean to control
whether or not children, e.g. ListItemButtons, fire the composite's
on_release, via a call to trigger_action().

Removed similar methods from adapters, so the adapters, and selection,
react, and are not as involved in the underlying machinery of how
selection events are fired.

The select() and deselect() methods for SelectableView and subclasses
now do both cosmetic UI changes and the setting of is_selected.

Updated the examples and tests for these changes.