Adaptergeddon complete; ListView using controllers
A single remnant of adapters, as far as named files go, is kivy.adapter,
which is the base class for ListView.

The code of adapters was reused in ListController, DictController, and
in ListView. Internally, the arrangement of detailed op dispatching from
OpObservableList and Dict to data_changed() methods remains, but now
this code is in ListController and ListView.

There is a new binding system used by controllers and ListView that
works in python and in kv, which was a primary goal. To make this work
in kv, a bindings lists is passed through the recursive apply machinery
of lang.py, where all bindings are accumulated and associated with their
widgets. Although to make this work in kv, the Binding base class was
changed to Widget, the bindings are sent to their widgets (ListView so
far), to a special init_kv_bindings() method. This is done because of
the timing and manner of instantation in kv vs. python. Perhaps not the
final solution, but now there is a fairly short syntax for defining a
ListView in kv, to use a given controller.

ListView examples were pared down and rewritten. ListView docs were also
reduced substantially, with some terse references to the examples that
can be fleshed out later.

Tests were given a quick once-over to delete the obviously invalid ones,
with adjustments made to make much of the ones added in this branch to
pass again. (The remaining ones pass, but some were deleted).

There will be continued work needed to clean up everthing and to fix a
few remaining known serious bugs:

1) There is something wrong with the end of data handling, which could
specifically have to do with scrolling, where it shows up as app locking
if you scroll to the end.

2) The list_cascade.py example uses a dynamic class (templates were
removed), but there is a first-use glitch where the root.text is not set
for the first item. This is in the area of kv widget instantiation, for
use of dynamic classes.

There are undoubtedly others, but it is important to bring this branch
out of the forest of refactoring, back into the mix. Additional work
will be needed, but the examples are back in working order, allowing the
(re)start of evaluation.